+++ src/components/elements/Input.tsx
@@ -1,6 +1,6 @@
 import type { UseFormRegisterReturn, FieldError } from 'react-hook-form';
 
-import { INPUT_CLASS } from '../../utils/constans/input-constans';
+import { INPUT_CLASS } from '../../utils/constans/inputConstans';
 
 type InputProps = {
   class: 'register' | 'search',
