+++ src/pages/NotFound.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
 
-import { BUTTON_CLASS } from '../utils/constans/button-constans';
+import { BUTTON_CLASS } from '../utils/constans/buttonConstans';
 
 function NotFound() {
   return (
