+++ src/service/slice/disciplineSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { IDisciplineState } from '../../utils/interface/disciplineInterface';
 import { disciplinesNames, disciplineContent } from '../actions/disciplineAction';
-import { DEFAULT_DISCIPLINE } from '../../utils/constans/default-discipline-constans';
+import { DEFAULT_DISCIPLINE } from '../../utils/constans/defaultDisciplineConstans';
 
 
 const initialState: IDisciplineState = {
